lots of update, I just cannot remember what I've written

+++ src/org/ice1000/bit/IntervalUpdatePointQuery.java
@@ -23,7 +23,7 @@ public final class IntervalUpdatePointQuery extends BinaryIndexedTree {
 	 * @param end   right bound of [begin, end] <= length
 	 * @param value every element in [begin, end] will 'plus assign' value.
 	 */
-	public void update(int begin, int end, int value) {
+	public void update(int begin, int end, long value) {
 		if (end < begin) throw new BinaryIndexedTreeException("end should be smaller than begin!");
 		add(begin, value);
 		add(end + 1, -value);
